Refactor local_scans, add predict_save_path, use_scalers toggle
Issues #16, #24, and follow-up readability work:

- Add CountersMixin.predict_save_path default (no-EPICS-I/O path
  prediction from hdf1_name_format/hdf1_file_format) and
  LightFieldDetector override for its different path logic.
- Add use_scalers bool to CountersClass; when False, unselected scalers
  are not auto-added to detectors (Issue #24).
- Extract private helpers from local_scans.py into new
  _local_scan_utils.py: LocalFlag/flag, _collect_extras, dichro_steps,
  one_local_step, one_local_shot, _setup_detectors, plus new helpers
  _configure_dichro, _configure_fixq, _default_per_step,
  _default_per_shot, _build_scan_md, _check_magnet911.
- Replace _setup_paths + setup_nxwritter with _setup_file_io: three-
  phase validate-before-configure approach so no detector PVs are
  written if any output file already exists (Issue #16).
- local_scans.py reduced from 1333 to ~900 lines; each plan preamble
  collapsed from ~50 to ~15 lines.
