+++ src/id4_common/plans/local_scans.py
@@ -15,16 +15,12 @@ __all__ = [
 ]
 
 from logging import getLogger
-from pathlib import Path
 
 from apsbits.core.instrument_init import oregistry
-from apsbits.utils.config_loaders import get_config
 from bluesky.plan_patterns import chunk_outer_product_args
 from bluesky.plan_stubs import abs_set as bps_abs_set
-from bluesky.plan_stubs import move_per_step
 from bluesky.plan_stubs import mv as bps_mv
 from bluesky.plan_stubs import rd
-from bluesky.plan_stubs import trigger_and_read
 from bluesky.plans import count as bp_count
 from bluesky.plans import grid_scan as bp_grid_scan
 from bluesky.plans import list_scan
@@ -38,312 +34,28 @@ from toolz import partition
 
 from ..callbacks.dichro_stream import dichro as dichro_device
 from ..callbacks.nexus_data_file_writer import nxwriter
-from ..utils.counters_class import counters
-from ..utils.experiment_utils import experiment
 from ..utils.hkl_utils import current_diffractometer
-from ..utils.run_engine import RE
+from ._local_scan_utils import _build_scan_md
+from ._local_scan_utils import _check_magnet911
+from ._local_scan_utils import _collect_extras
+from ._local_scan_utils import _configure_dichro
+from ._local_scan_utils import _configure_fixq
+from ._local_scan_utils import _default_per_shot
+from ._local_scan_utils import _default_per_step
+from ._local_scan_utils import _setup_detectors
+from ._local_scan_utils import _setup_file_io
+from ._local_scan_utils import flag
 from .local_preprocessors import configure_counts_decorator
 from .local_preprocessors import extra_devices_decorator
 from .local_preprocessors import stage_4idg_softglue_decorator
 from .local_preprocessors import stage_dichro_decorator
 from .local_preprocessors import stage_magnet911_decorator
 
-try:
-    # change to import this only if needed?
-    from ..utils.pr_setup import pr_setup
-except ModuleNotFoundError:
-    pr_setup = None
-
-iconfig = get_config()
-
 logger = getLogger(__name__)
 logger.info(__file__)
 
-HDF1_NAME_FORMAT = Path(iconfig["AREA_DETECTOR"]["HDF5_FILE_TEMPLATE"])
-
-
-class LocalFlag:
-    """Stores flags that are used to select and run local scans."""
-
-    dichro = False
-    fixq = False
-    hkl_pos = {}
-    dichro_steps = None
-    vortex_sgz = False
-
-
-flag = LocalFlag()
-
-
-def _collect_extras(args):
-    """Collect all detectors that need to be read during a scan."""
-
-    # TODO: most or all of this can be removed if we add these to the energy
-    # device directly.dichro_bec
-
-    # Initialize the list of extra devices with the standard set from counters
-    extras = counters.extra_devices.copy()
-
-    energy = oregistry.find("energy", allow_none=True)
-    escan_flag = False if energy is None or energy not in args else True
-    if escan_flag:
-        undulators = oregistry.find("undulators", allow_none=True)
-        if undulators is None:
-            logger.warning(
-                "Undulators device not found. Will not record the undulator "
-                "energy during scan."
-            )
-        else:
-            for und in (undulators.ds, undulators.us):
-                und_track = yield from rd(und.tracking)
-                if und_track:
-                    extras.append(und.energy)
-
-        # Do the same for phase plates
-        prs = oregistry.findall("phase retarders", allow_none=True)
-        if prs is None:
-            logger.warning(
-                "No phase retarder device not found. Will not record its "
-                "position energy during scan."
-            )
-        else:
-            for pr in prs:
-                # Fetch tracking status asynchronously
-                pr_track = yield from rd(pr.tracking)
-                if pr_track:
-                    extras.append(pr.th)
-
-    diff = current_diffractometer()
-    huber_flag = False if diff is None or diff.name not in str(args) else True
-    if huber_flag:
-        extras.append(current_diffractometer())
-
-    return extras
-
-
-def dichro_steps(devices_to_read, take_reading):
-    """
-    Switch the x-ray polarization for each scan point.
-    This will increase the number of points in a scan by a factor that is equal
-    to the length of the `pr_setup.dichro_steps` list.
-    """
-    devices_to_read += [pr_setup.positioner]
-    center = 0 if pr_setup.oscillate_pzt else pr_setup.positioner.position
-
-    for pos in flag.dichro_steps:
-        yield from mv(pr_setup.positioner, pos + center)
-        yield from take_reading(devices_to_read)
-
-
-def one_local_step(detectors, step, pos_cache, take_reading=trigger_and_read):
-    """
-    Inner loop for fixQ and dichro scans.
-
-    It is always called in the local plans defined here. It is used as a
-    `per_step` kwarg in Bluesky scan plans, such as `bluesky.plans.scan`. But
-    note that it requires the `LocalFlag` class.
-
-    Parameters
-    ----------
-    detectors : iterable
-        devices to read
-    step : dict
-        mapping motors to positions in this step
-    pos_cache : dict
-        mapping motors to their last-set positions
-    take_reading : plan, optional
-        function to do the actual acquisition ::
-           def take_reading(dets, name='primary'):
-                yield from ...
-        Callable[List[OphydObj], Optional[str]] -> Generator[Msg], optional
-        Defaults to `trigger_and_read`
-    """
-
-    devices_to_read = list(step.keys()) + list(detectors)
-    yield from move_per_step(step, pos_cache)
-
-    if flag.fixq:
-        huber = current_diffractometer()
-        devices_to_read += [huber]
-        args = (
-            huber.h,
-            flag.hkl_pos[huber.h],
-            huber.k,
-            flag.hkl_pos[huber.k],
-            huber.l,
-            flag.hkl_pos[huber.l],
-        )
-        yield from bps_mv(*args)
-
-    if flag.dichro:
-        yield from dichro_steps(devices_to_read, take_reading)
-    else:
-        if flag.vortex_sgz:
-            # TODO: Is there a better way for this?
-            sgz_vortex = oregistry.find("sgz_vortex")
-            vortex = oregistry.find("vortex")
-
-            # Reset SGZ
-            yield from sgz_vortex.reset()
-
-            # Arm Vortex
-            yield from vortex.arm_plan()
-            # yield from sleep(vortex._sleep_after_trigger)
-            # logger.debug(vortex._sleep_after_trigger)
-
-        yield from take_reading(devices_to_read)
-
-
-def one_local_shot(detectors, take_reading=trigger_and_read):
-    """
-    Inner loop for fixQ and dichro scans.
-    To be used as a `per_shot` kwarg in the Bluesky `bluesky.plans.count`.
-    It is always called in the local `count` plan defined here. It is used as a
-    `per_shot` kwarg in the Bluesky `bluesky.plans.count`. But note that it
-    requires the `LocalFlag` class.
-    Parameters
-    ----------
-    detectors : iterable
-        devices to read
-    take_reading : plan, optional
-        function to do the actual acquisition ::
-           def take_reading(dets, name='primary'):
-                yield from ...
-        Callable[List[OphydObj], Optional[str]] -> Generator[Msg], optional
-        Defaults to `trigger_and_read`
-    """
-
-    devices_to_read = list(detectors)
-    if flag.dichro:
-        yield from dichro_steps(devices_to_read, take_reading)
-    else:
-        if flag.vortex_sgz:
-            # TODO: Is there a better way for this?
-            sgz_vortex = oregistry.find("sgz_vortex")
-            vortex = oregistry.find("vortex")
-
-            # Reset SGZ
-            yield from sgz_vortex.reset()
-
-            # Arm Vortex
-            yield from vortex.arm_plan()
-
-        yield from take_reading(devices_to_read)
-
-
-def _setup_paths(detectors):
-    if None in (experiment.base_experiment_path, experiment.file_base_name):
-        raise ValueError(
-            "The experiment needs to be setup, please run experiment_setup()"
-        )
-
-    _scan_id = RE.md["scan_id"] + 1
-
-    # Master file
-    _master_fullpath = str(HDF1_NAME_FORMAT) % (
-        str(experiment.experiment_path),
-        experiment.file_base_name,
-        _scan_id,
-    )
-    _master_fullpath += "_master.hdf"
-
-    # Setup area detectors
-    _dets_file_paths = {}
-    # Relative paths are used in the master file so that data can be copied.
-    _rel_dets_paths = {}
-    for det in list(detectors):
-        # Check if we can and want to get images from this detector
-        _setup_images = getattr(det, "setup_images", None)
-        _flag = getattr(det, "save_image_flag", False)
-        if _setup_images and _flag:
-            _fp, _rp = _setup_images(
-                experiment.experiment_path,
-                experiment.file_base_name,
-                _scan_id,
-                flyscan=False,
-            )
-            _dets_file_paths[det.name] = str(_fp)
-            _rel_dets_paths[det.name] = str(_rp)
-
-    # Check if any of these files exists
-    for _fname in [_master_fullpath] + list(_dets_file_paths.values()):
-        if Path(_fname).is_file():
-            raise FileExistsError(
-                f"The file {_fname} already exists! Will not overwrite, quitting."
-            )
-
-    return _master_fullpath, _dets_file_paths, _rel_dets_paths
-
-
-def setup_nxwritter(_base_path, _master_fullpath, _rel_dets_paths):
-    nxwriter.external_files = _rel_dets_paths
-    nxwriter.file_name = str(_master_fullpath)
-    nxwriter.file_path = str(_base_path)
-
-
-def setup_detectors(time):
-    # If counting against time, then all good, can just use the detectors
-    # setup in counters
-    if time > 0:
-        dets = counters.detectors
-
-        if flag.vortex_sgz:
-            vortex = oregistry.find("vortex", allow_none=True)
-            if vortex is None:
-                raise ValueError(
-                    "Vortex detector not found by oregistry! It is "
-                    "required for vortex_sgz mode."
-                )
-
-            sgz_vortex = oregistry.find("sgz_vortex", allow_none=True)
-            if sgz_vortex is None:
-                raise ValueError(
-                    "sgz_vortex detector not found by oregistry! It is "
-                    "required for vortex_sgz mode."
-                )
-            if vortex not in dets:
-                dets.append(vortex)
-            if sgz_vortex not in dets:
-                dets.append(sgz_vortex)
-
-            vortex._vortex_sgz = True
-
-            freq = 1 / sgz_vortex.div_by_n.n.get() * sgz_vortex.clock_freq.get()
-            pulses = int(time * freq)
-            sgz_vortex.down_counter_pulse.preset.set(pulses).wait(5)
-
-        return dets  # this should have all scalers by default
-    # If counting against a monitor, it only works if the monitor is a scaler
-    # channel and all detectors are in the same scaler.
-    else:
-        if not counters.is_scaler_monitor:
-            return counters.detectors
-
-        if counters.monitor == "Time":
-            raise ValueError(
-                "Monitor is set to 'Time', but you are trying to count against "
-                "a scaler channel. Please run counters.plotselect() to change "
-                "the monitor to a scaler channel."
-            )
-
-        monitor_scaler = counters.detectors_plot_options.loc[
-            counters.detectors_plot_options["channels"] == counters.monitor
-        ]["detectors"].iloc[0]
-
-        if any(
-            [
-                det_name != monitor_scaler
-                for det_name in counters.selected_plot_detectors
-            ]
-        ):
-            raise ValueError(
-                "You can only count against monitor if all detectors are in "
-                "the same scaler of the monitor. But "
-                f"{counters.selected_plot_detectors} have been selected."
-                "Run counters.plotselect() to change the detector channels."
-            )
-
-        return [oregistry.find(monitor_scaler)]
+# Keep backward-compatible alias
+setup_detectors = _setup_detectors
 
 
 def count(
@@ -360,8 +72,10 @@ def count(
 ):
     """
     Take one or more readings from detectors.
+
     This is a local version of `bluesky.plans.count`. Note that the `per_shot`
     cannot be set here, as it is used for dichro scans.
+
     Parameters
     ----------
     num : integer
@@ -393,6 +107,7 @@ def count(
         for details.
     md : dict, optional
         metadata
+
     Notes
     -----
     If ``delay`` is an iterable, it must have at least ``num - 1`` entries or
@@ -407,58 +122,28 @@ def count(
 
     flag.vortex_sgz = vortex_sgz
 
-    fixq = False
     if detectors is None:
-        detectors = setup_detectors(time)
+        detectors = _setup_detectors(time)
 
-    flag.dichro = dichro
-    if dichro:
-        _offset = pr_setup.offset.get()
+    _configure_dichro(dichro)
+    _configure_fixq(False)
 
-        if pr_setup.oscillate_pzt:
-            _center = pr_setup.positioner.parent.center.get()
-        else:
-            _center = -1 * _offset
-
-        _steps = pr_setup.dichro_steps
-        flag.dichro_steps = [_center + step * _offset for step in _steps]
-
-    flag.fixq = fixq
-
-    if per_shot is not None and (fixq or dichro):
-        logger.warning(
-            "there is a custom per_shot, but fixQ or dichro was selected."
-        )
+    if per_shot is not None and dichro:
+        logger.warning("there is a custom per_shot, but dichro was selected.")
     elif per_shot is None:
-        per_shot = one_local_shot if (fixq or dichro or vortex_sgz) else None
+        per_shot = _default_per_shot(dichro, vortex_sgz)
 
-    _master_fullpath, _dets_file_paths, _rel_dets_paths = _setup_paths(
+    _master_fullpath, _dets_file_paths, _rel_dets_paths = _setup_file_io(
         detectors if not g_sgz else detectors + [pos_stream]
-    )
-
-    setup_nxwritter(
-        experiment.experiment_path, _master_fullpath, _rel_dets_paths
     )
 
     extras = yield from _collect_extras(("",))
 
-    # TODO: The md handling might go well in a decorator.
-    # TODO: May need to add reference to stream.
-    _md = dict(
-        hints={"monitor": counters.monitor_field, "detectors": []},
-        data_management=experiment.data_management or "None",
-        esaf=experiment.esaf,
-        proposal=experiment.proposal,
-        base_experiment_path=str(experiment.base_experiment_path),
-        experiment_path=str(experiment.experiment_path),
-        master_file_path=str(_master_fullpath),
-        detectors_file_full_path=_dets_file_paths,
-        detectors_file_relative_path=_rel_dets_paths,
+    _md = _build_scan_md(
+        detectors, _master_fullpath, _dets_file_paths, _rel_dets_paths
     )
-
     for item in detectors:
         _md["hints"]["detectors"].extend(item.hints["fields"])
-
     _md.update(md or {})
 
     @stage_magnet911_decorator(False)
@@ -555,65 +240,30 @@ def ascan(
     flag.vortex_sgz = vortex_sgz
 
     if detectors is None:
-        # detectors = counters.detectors
-        detectors = setup_detectors(time)
+        detectors = _setup_detectors(time)
 
-    flag.dichro = dichro
-    if dichro:
-        _offset = pr_setup.offset.get()
+    _configure_dichro(dichro)
+    _configure_fixq(fixq)
 
-        if pr_setup.oscillate_pzt:
-            _center = pr_setup.positioner.parent.center.get()
-        else:
-            _center = -1 * _offset
+    per_step = per_step or _default_per_step(fixq, dichro, vortex_sgz)
 
-        _steps = pr_setup.dichro_steps
-        flag.dichro_steps = [_center + step * _offset for step in _steps]
-
-    flag.fixq = fixq
-    if per_step is None:
-        per_step = one_local_step if (fixq or dichro or vortex_sgz) else None
-    if fixq:
-        huber = current_diffractometer()
-        flag.hkl_pos = {
-            huber.h: huber.h.get().setpoint,
-            huber.k: huber.k.get().setpoint,
-            huber.l: huber.l.get().setpoint,
-        }
-
-    _master_fullpath, _dets_file_paths, _rel_dets_paths = _setup_paths(
+    _master_fullpath, _dets_file_paths, _rel_dets_paths = _setup_file_io(
         detectors if not g_sgz else detectors + [pos_stream]
-    )
-
-    setup_nxwritter(
-        experiment.experiment_path, _master_fullpath, _rel_dets_paths
     )
 
     extras = yield from _collect_extras(args)
 
-    _md = dict(
-        hints={"monitor": counters.monitor_field, "detectors": []},
-        data_management=experiment.data_management or "None",
-        esaf=experiment.esaf,
-        proposal=experiment.proposal,
-        base_experiment_path=str(experiment.base_experiment_path),
-        experiment_path=str(experiment.experiment_path),
-        master_file_path=str(_master_fullpath),
-        detectors_file_full_path=_dets_file_paths,
-        detectors_file_relative_path=_rel_dets_paths,
+    _md = _build_scan_md(
+        detectors, _master_fullpath, _dets_file_paths, _rel_dets_paths
     )
-
     for item in detectors:
         _md["hints"]["detectors"].extend(item.hints["fields"])
-
     _md["hints"]["scan_type"] = "ascan"
-
     _md.update(md or {})
 
     motors = [motor for motor, _, _ in partition(3, args)]
 
-    magnet911 = oregistry.find("magnet911", allow_none=True)
-    magnet_option = False if magnet911 is None else (magnet911.ps.field in args)
+    magnet_option = _check_magnet911(args)
 
     @stage_magnet911_decorator(magnet_option)
     @stage_4idg_softglue_decorator(g_sgz)
@@ -624,7 +274,6 @@ def ascan(
     @subs_decorator(nxwriter.receiver)
     def _inner_ascan():
         yield from scan(detectors + extras, *args, per_step=per_step, md=_md)
-
         yield from nxwriter.wait_writer_plan_stub()
 
     return (yield from _inner_ascan())
@@ -676,7 +325,7 @@ def lup(
         Note that hkl is moved ~after~ the other motors!
     vortex_sgz : boolean, optional
         Measures the Vortex detector using the softgluezynq triggers. This is a
-        special mode that requires the 'vortex' and 'sgz_vortex' devices to\
+        special mode that requires the 'vortex' and 'sgz_vortex' devices to
         exist otherwise an error will be thrown.
     per_step: callable, optional
         hook for customizing action of inner loop (messages per step).
@@ -821,6 +470,7 @@ def grid_scan(
 ):
     """
     Scan over a mesh; each motor is on an independent trajectory.
+
     Parameters
     ----------
     ``*args``
@@ -891,65 +541,30 @@ def grid_scan(
     flag.vortex_sgz = vortex_sgz
 
     if detectors is None:
-        detectors = setup_detectors(time)
+        detectors = _setup_detectors(time)
 
-    flag.dichro = dichro
-    if dichro:
-        _offset = pr_setup.offset.get()
+    _configure_dichro(dichro)
+    _configure_fixq(fixq)
 
-        if pr_setup.oscillate_pzt:
-            _center = pr_setup.positioner.parent.center.get()
-        else:
-            _center = -1 * _offset
+    per_step = per_step or _default_per_step(fixq, dichro, vortex_sgz)
 
-        _steps = pr_setup.dichro_steps
-        flag.dichro_steps = [_center + step * _offset for step in _steps]
-
-    flag.fixq = fixq
-    if per_step is None:
-        per_step = one_local_step if (fixq or dichro or vortex_sgz) else None
-
-    if fixq:
-        huber = current_diffractometer()
-        flag.hkl_pos = {
-            huber.h: huber.h.get().setpoint,
-            huber.k: huber.k.get().setpoint,
-            huber.l: huber.l.get().setpoint,
-        }
-
-    _master_fullpath, _dets_file_paths, _rel_dets_paths = _setup_paths(
+    _master_fullpath, _dets_file_paths, _rel_dets_paths = _setup_file_io(
         detectors if not g_sgz else detectors + [pos_stream]
-    )
-
-    setup_nxwritter(
-        experiment.experiment_path, _master_fullpath, _rel_dets_paths
     )
 
     extras = yield from _collect_extras(args)
 
-    _md = dict(
-        hints={"monitor": counters.monitor_field, "detectors": []},
-        data_management=experiment.data_management or "None",
-        esaf=experiment.esaf,
-        proposal=experiment.proposal,
-        base_experiment_path=str(experiment.base_experiment_path),
-        experiment_path=str(experiment.experiment_path),
-        master_file_path=str(_master_fullpath),
-        detectors_file_full_path=_dets_file_paths,
-        detectors_file_relative_path=_rel_dets_paths,
+    _md = _build_scan_md(
+        detectors, _master_fullpath, _dets_file_paths, _rel_dets_paths
     )
-
     for item in detectors:
         _md["hints"]["detectors"].extend(item.hints["fields"])
-
     _md["hints"]["scan_type"] = "gridscan"
-
     _md.update(md or {})
 
     motors = [m[0] for m in chunk_outer_product_args(args)]
 
-    magnet911 = oregistry.find("magnet911", allow_none=True)
-    magnet_option = False if magnet911 is None else (magnet911.ps.field in args)
+    magnet_option = _check_magnet911(args)
 
     @stage_magnet911_decorator(magnet_option)
     @stage_4idg_softglue_decorator(g_sgz)
@@ -966,7 +581,6 @@ def grid_scan(
             per_step=per_step,
             md=_md,
         )
-
         yield from nxwriter.wait_writer_plan_stub()
 
     return (yield from _inner_grid_scan())
@@ -1123,34 +737,12 @@ def qxscan(
     flag.vortex_sgz = vortex_sgz
 
     if detectors is None:
-        detectors = setup_detectors(time)
+        detectors = _setup_detectors(time)
 
-    flag.dichro = dichro
-    if dichro:
-        _offset = pr_setup.offset.get()
+    _configure_dichro(dichro)
+    _configure_fixq(fixq)
 
-        if pr_setup.oscillate_pzt:
-            _center = pr_setup.positioner.parent.center.get()
-        else:
-            _center = -1 * _offset
-
-        _steps = pr_setup.dichro_steps
-        flag.dichro_steps = [_center + step * _offset for step in _steps]
-
-    flag.fixq = fixq
-
-    if per_step is None:
-        per_step = one_local_step if (fixq or dichro or vortex_sgz) else None
-
-    if fixq:
-        huber = current_diffractometer()
-        flag.hkl_pos = {
-            huber.h: huber.h.get().setpoint,
-            huber.k: huber.k.get().setpoint,
-            huber.l: huber.l.get().setpoint,
-        }
-
-    # Get energy argument and extras
+    per_step = per_step or _default_per_step(fixq, dichro, vortex_sgz)
 
     energy = oregistry.find("energy", allow_none=True)
     if energy is None:
@@ -1169,43 +761,24 @@ def qxscan(
     factor_list = yield from rd(qxscan_setup.factor_list)
 
     _ct = {}
-
     for det in detectors:
         _ct[det] = abs(time)
         args += (det.preset_monitor, abs(time) * array(factor_list))
 
-    _master_fullpath, _dets_file_paths, _rel_dets_paths = _setup_paths(
+    _master_fullpath, _dets_file_paths, _rel_dets_paths = _setup_file_io(
         detectors if not g_sgz else detectors + [pos_stream]
     )
 
-    setup_nxwritter(
-        experiment.experiment_path, _master_fullpath, _rel_dets_paths
+    _md = _build_scan_md(
+        detectors, _master_fullpath, _dets_file_paths, _rel_dets_paths
     )
-
-    _md = dict(
-        hints={"monitor": counters.monitor_field, "detectors": []},
-        data_management=experiment.data_management or "None",
-        esaf=experiment.esaf,
-        proposal=experiment.proposal,
-        base_experiment_path=str(experiment.base_experiment_path),
-        experiment_path=str(experiment.experiment_path),
-        master_file_path=str(_master_fullpath),
-        detectors_file_full_path=_dets_file_paths,
-        detectors_file_relative_path=_rel_dets_paths,
-    )
-
-    # TODO: The md handling might go well in a decorator.
-    # TODO: May need to add reference to stream.
-    # _md = {'hints': {'monitor': counters.monitor, 'detectors': []}}
     for item in detectors:
         _md["hints"]["detectors"].extend(item.hints["fields"])
-
     _md["hints"]["scan_type"] = "qxscan"
     if dichro:
         _md["hints"]["scan_type"] += " dichro"
     if lockin:
         _md["hints"]["scan_type"] += " lockin"
-
     _md.update(md or {})
 
     @monitor_during_decorator([dichro_device] if dichro else [])
@@ -1218,7 +791,6 @@ def qxscan(
         yield from list_scan(
             detectors + extras, *args, per_step=per_step, md=_md
         )
-
         # put original times back.
         for det, preset in _ct.items():
             yield from mv(det.preset_monitor, preset)
@@ -1251,8 +823,7 @@ def mv(*args, **kwargs):
     :func:`bluesky.plan_stubs.mv`
     """
 
-    magnet911 = oregistry.find("magnet911", allow_none=True)
-    magnet_option = False if magnet911 is None else (magnet911.ps.field in args)
+    magnet_option = _check_magnet911(args)
 
     @stage_magnet911_decorator(magnet_option)
     def _inner_mv():
@@ -1264,6 +835,7 @@ def mv(*args, **kwargs):
 def mvr(*args, **kwargs):
     """
     Move one or more devices to a relative setpoint. Wait for all to complete.
+
     If more than one device is specified, the movements are done in parallel.
 
     This is a local version of `bluesky.plan_stubs.mvr`.
@@ -1274,9 +846,11 @@ def mvr(*args, **kwargs):
         device1, value1, device2, value2, ...
     kwargs :
         passed to bluesky.plan_stub.mvr
+
     Yields
     ------
     msg : Msg
+
     See Also
     --------
     :func:`bluesky.plan_stubs.rel_set`
@@ -1296,6 +870,7 @@ def mvr(*args, **kwargs):
 def abs_set(*args, **kwargs):
     """
     Set a value. Optionally, wait for it to complete before continuing.
+
     This is a local version of `bluesky.plan_stubs.abs_set`. If more than one
     device is specifed, the movements are done in parallel.
 
@@ -1323,8 +898,7 @@ def abs_set(*args, **kwargs):
     :func:`bluesky.plan_stubs.mv`
     """
 
-    magnet911 = oregistry.find("magnet911", allow_none=True)
-    magnet_option = False if magnet911 is None else (magnet911.ps.field in args)
+    magnet_option = _check_magnet911(args)
 
     @stage_magnet911_decorator(magnet_option, persistent=False)
     def _inner_abs_set():
